### 2024-09-27 #7296 @editorialbot
[PRE REVIEW]: GraphBin-Tk: assembly graph-based metagenomic binning toolkit
**Submitting author:** @Vini2 (<a href="http://orcid.org/0000-0002-2651-8719">Vijini Mallawaarachchi</a>)
**Repository:** https://github.com/metagentools/gbintk
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
