### 2025-01-20 #7296 @arfon
:wave: @vinisalazar @telatin @cumbof @KatyBrown @aboffin – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _GraphBin-Tk: assembly graph-based metagenomic binning toolkit_

Many thanks
Arfon
