### 2025-02-11 #7713 @vinisalazar
The authors present a Python toolkit, "GraphBin-Tk", which facilitates the execution of a suite of metagenomic binning tools previously published by the authors, namely "MetaCoAg", "GraphBin", and "GraphBin2". This toolkit wraps the execution of these software under a user-friendly CLI tool. Moreover, it provides commands for evaluating and visualising the results of said tools.

This appears to be a high-quality submission and within the scope set out by JOSS. I have, however, noted a few concerns:

- The "Statement of Need" section bridges directly into explanation of each of the subcommands, and thus comprises the main section of the paper. This section highlights some of the challenges related to the task of metagenomic binning, but deviates from the main focus of the software, which is to facilitate the execution of this procedure, that is rather complex and may require multiple steps. I believe that this section can be revised to be more concise, and more clearly highlight the need for "gbintk". It then should go on to a separate section that describes each of the subcommands and other relevant results.
- I was keen to run the software myself, but I was unable to find the Sim-5G, Sim-10G and Sim-20G datasets on the linked references (V. G. Mallawaarachchi et al. 2020, 2021). The supplementary material provides some information on the datasets (ie. how they were generated), but no data files.
-  The results only compare the results of MetaCoAg, and MetaCoAg + GraphBin. It does not seems to include any performance information regarding GraphBin2 (perhaps an extra bar can be included on Figure 3 showing this?).
- It would be helpful to list other metagenomic binning wrappers/toolkits, e.g. [MetaWRAP](https://github.com/bxlab/metaWRAP) or [DAS_Tool](https://github.com/cmks/DAS_Tool).
- The README section mentions that GraphBin-Tk depends on MetaCoAg, GraphBin, and GraphBin2, and these tools must be cited. These tools themselves, however, rely on a number of third-party software, e.g. Cogent, SciPy, Python-igraph. It should direct users to cite these tools, as well as any assemblers used to generate the assembly graph (e.g. MEGAHIT).
- Minor: there's a typo on the word "intiial" in the caption of Figure 3.

These concerns have prevented me from checking the missing boxes in the supplied checklist. Once they're addressed, I would be glad to take another look.

Regards,
V
