### 2025-02-24 metagentools/gbintk#14 @vinisalazar
JOSS review
Please see https://github.com/openjournals/joss-reviews/issues/7713#issuecomment-2649523813 for the review.
