### 2025-02-25 metagentools/gbintk#15 @telatin
JOSS review tracker
https://github.com/openjournals/joss-reviews/issues/7713
