### 2025-03-07 metagentools/gbintk#16 @cumbof
JOSS Review
This issue is part of the review process for your submission to JOSS [openjournals/joss-reviews/issues/7713](https://github.com/openjournals/joss-reviews/issues/7713).
Please, do not close this issue until all the following points have been addressed. Keep in mind that this is a WIP issue, so I may keep extending the following list with new comments until I will finally close the issue.

1. The installation of `gbintk` is straightforward, but the [documentation](https://gbintk.readthedocs.io/en/latest/install/) needs to be expanded with information about software dependencies (both python packages and external software requirements).

2. The recommended Python version must also be clearly highlighted in the documentation. What if I install your software via `pip`? I may need to manually install all the external software dependencies like `MetaCoAG`, `GraphBin`, `GraphBin2`, `SPAdes`, `MEGAHIT`, etc, so that must be clearly stated in the documentation;

3. The _Example Usage_ section of [MetaCoAG](https://gbintk.readthedocs.io/en/latest/metacoag_usage/#example-usage), [GraphBin](https://gbintk.readthedocs.io/en/latest/graphbin_usage/#example-usage), [GraphBin2](https://gbintk.readthedocs.io/en/latest/graphbin2_usage/#example-usage), [Prepare](https://gbintk.readthedocs.io/en/latest/prepare/#example-usage), [Visualise](https://gbintk.readthedocs.io/en/latest/visualise/#example-usage), and [Evaluate](https://gbintk.readthedocs.io/en/latest/evaluate/#example-usage) must be more informative than providing a simple command line example with fake paths to some kind of data. Please provide real examples with real data, like you did with your `pytest` [tests](https://github.com/metagentools/gbintk/tree/main/tests) and [DEMO.md](https://github.com/metagentools/gbintk/blob/main/DEMO.md);

4. In your manuscript, you used the _Sim-5G+metaSPAdes_ dataset as an example for your figures. It is great that you cited _V. G. Mallawaarachchi et al., 2020, 2021_, but please also add a direct link to the dataset, preferably hosted on Zenodo;

5. Please correct the Figure 3 caption in your manuscript (__intiial__ to __initial__);
