### 2025-04-04 metagentools/gbintk#15 @telatin
I checked the tool and found no issues.

It was a bit hard to grasp the use case and I would recommend:

1. To remove usage instructions from the paper (they are better suited for README and other docs as they will also evolve). Instead, please add more background on where the toolkit sit in the broader metagenomics workflow, and what need it addresses for people already using the tools wrapped in it.

2. In parallel, please expand the docs to better describe what upstream steps are required to run the subcommand.

No other comments from me, nice work :)
