### 2025-04-22 d66fe6a
DEV: update pyproject.toml
