### 2025-04-22 f3d0511
DEV: update pyproject.toml
