### 2025-04-22 1196f57
DEV: revert pyproject.toml
