### 2025-04-22 e81673c
DEV: update pyproject.toml
