+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = ["click",
   "cogent3",
   "igraph",
   "cairocffi",
-  "pycairo",
+  "pycairo<=1.27.0",
   "networkx",
   "scipy",
   "numpy",
