+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = ["click",
   "cogent3",
   "igraph",
   "cairocffi",
-  "pycairo<=1.27.0",
+  "pycairo",
   "networkx",
   "scipy",
   "numpy",
