### 2025-04-24 e8c806a
DOC: update citations in README
