### 2025-04-24 #7713 @telatin
Yes, to me the main (suggested) changes are in the paper itself (https://github.com/metagentools/gbintk/issues/15#issuecomment-2777885989)
