+++ paper/paper.md
@@ -46,155 +46,39 @@ Conventional metagenomic binning tools make use of features such as nucleotide c
 
 # Statement of need
 
-It is crucial to obtain accurate binning results in metagenomic studies to understand the composition and functional potential of microbial communities. Conventional binning methods mainly rely on two features of contigs; 1) nucleotide composition, represented as normalised frequencies of oligonucleotides (short substrings of a particular length) and 2) abundance, the average number of reads that covers each nucleotide base of the contig [@Mallawaarachchi:2024]. Previous studies have found that these tools face several challenges when binning complex datasets [@Mallawaarachchi:2024]:
+It is crucial to obtain accurate binning results in metagenomic studies to understand the composition and functional potential of microbial communities. Conventional binning methods mainly rely on two features of contigs; 1) nucleotide composition, represented as normalised frequencies of oligonucleotides (short substrings of a particular length) and 2) abundance, the average number of reads that covers each nucleotide base of the contig [@Mallawaarachchi:2024]. Previous studies have found that these tools face several challenges when binning complex datasets, especially those containing closely related species, short sequences, and shared genomic regions [@Mallawaarachchi:2024].
 
-* Contigs of closely related species can be merged, producing contaminated bins.
-* Contigs having features that deviate from their constituent genomes (e.g., the presence of protein-coding regions and repeat regions can lead to deviated nucleotide and abundance features) can be placed in incorrect bins.
-* Contigs that are too short (e.g., shorter than 1000 base pairs) can be discarded during binning as they may not capture enough genomic signatures. Such short sequences can contain important regions such as repeats.
-* Contigs shared among different genomes are only placed in the bin of the most representative genome.
+To address these challenges, several graph-based metagenomic binning tools [@Mallawaarachchi:2024] such as MetaCoAG [@Mallawaarachchi1:2022; @Mallawaarachchi2:2022] and GraphMB [@Lamurias:2022], and bin refinement tools such as GraphBin [@Mallawaarachchi1:2020] and GraphBin2 [@Mallawaarachchi2:2020; @Mallawaarachchi:2021] have been developed. These tools enhance the binning results by leveraging the connectivity information of the assembly graph -- an intermediate output generated from the assembly process, that is often discarded during downstream analysis. However, running these tools individually can be challenging. Users have to install and configure multiple software packages, resolve dependencies and manage different file formats, which raises the risk of errors. Furthermore, existing metagenomic binning toolkits and wrappers such as MetaWRAP [@Uritskiy:2018], DAS Tool [@Sieber:2018] and MetaBinner [@Wang:2023] aim to simplify metagenomic binning workflows, but they do not incorporate graph-based binning tools.
 
[…]
