+++ paper/paper.bib
@@ -485,3 +485,10 @@ doi={10.1186/s13059-022-02832-6},
 url={https://doi.org/10.1186/s13059-022-02832-6}
 }
 
+@misc{deepthought_hpc,
+  title        = {DeepThought HPC},
+  author       = {{Flinders University}},
+  year         = {2024},
+  howpublished = {\url{https://doi.org/10.25957/FLINDERS.HPC.DEEPTHOUGHT}},
+  note         = {High Performance Computing resource at Flinders University},
+}
