### 2025-04-29 d7b7dca
DOC: update evaluate docs
