+++ docs/evaluate.md
@@ -1,8 +1,8 @@
 ## Evaluating binning results
 
-You can use the `evaluate` subcommand to evaluate your binning results given a ground truth.
+You can use the `evaluate` subcommand to evaluate your binning results given a ground truth. This evaluation is possible only for simulated or mock metagenomes where the ground truth genomes of contigs are known. 
 
-Run `gbintk evaluate --help` or `gbintk evaluate -h` to list the help message for evaluation.
+You can run `gbintk evaluate --help` or `gbintk evaluate -h` to list the help message for evaluation.
 
 ```shell
 Usage: gbintk evaluate [OPTIONS]
@@ -21,28 +21,40 @@ Options:
   -h, --help               Show this message and exit.
 ```
 
+### Evaluation Metrics
+
+GraphBin-Tk uses the four common metrics 1) precision, 2) recall, 3) F1-score and 4) Adjusted Rand Index (ARI) that have been used in previous binning studies. These metrics are calculated as follows. The binning result is denoted as a $K \times S$ matrix with $K$ number of bins and $S$ number of ground truth taxa. In this matrix, the element $a_{ks}$ denotes the number of contigs binned to the $k^{th}$ bin and belongs to the $s^{th}$ taxa. $U$ denotes the number of unbinned contigs and $N$ denotes the total number of contigs. Following are the equations used to calculate the evaluation metrics.
+
+__Precision__ = $\frac{\sum_{k}max_s \{a_{ks}\}}{\sum_{k}\sum_{s}a_{ks}}$
+
+__Recall__ = $\frac{\sum_{s}max_k \{a_{ks}\}}{(\sum_{k}\sum_{s}a_{ks}+U)}$
+
+__F1-score__ = $2 \times \frac{Precision\times Recall}{Precision+Recall}$
+
+__ARI__ = $\frac{\sum_{k,s}\binom{a_{ks}}{2}-t_3}{\frac{1}{2}(t_1+t_2)-t_3}$ $where\;t_1 = \sum_{k}\binom{\sum_{s}a_{ks}}{2},\;t_2 = \sum_{s}\binom{\sum_{k}a_{ks}}{2},\; and\; t_3 = \frac{t_1t_2}{\binom{N}{2}}$ 
+
+Please refer to the supplementary material of the [GraphBin publication](https://doi.org/10.1093/bioinformatics/btaa180) for further details.
+
 ### Input Format
 
-`evaluate` subcommand takes in 2 files as inputs.
+The following inputs are required to run the `evaluate` subcommand.
 
-* Binning result containing the comma separated records of `contig id,bin number` (in `.csv` format)
-* Ground truth annotations containing the comma separated records of `contig id, groud truth bin` (in `.csv` format)
+* A delimited text file containing the ground truth (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
+* A delimited text file containing the binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
 
 ### Example Usage
 
 ```shell
-gbintk evaluate --binned /path/to/binning_res.csv --groundtruth /path/to/grouhdtruth.csv --output /path/to/output_folder
+gbintk evaluate --binned Sim-20G/graphbin_output.csv --groundtruth Sim-20G/ground_truth.csv --output Sim-20G/evaluation_results
 ```
 
 ### Output
 
-You will get a file named `evaluation_results.txt` that contains the confusion matrix and the evaluation metrics precition, recall, F1-score and adjusted rand index (ARI). 
-
[…]
