+++ docs/visualise.md
@@ -1,6 +1,6 @@
 # Visualising binning results
 
-You can use the `visualise` subcommand to visualise the binning results by denoting coloured contigs in the assembly graph according to their corresponding bins. You can visualise the initial binning result obtained from an existing binning tool and the final binning result obtained from GraphBin/GraphBin2 and compare.
+You can use the `visualise` subcommand to visualise the binning results by denoting coloured contigs in the assembly graph according to their corresponding bins. You can visualise the initial binning result obtained from an existing binning tool and the final binning result obtained from GraphBin/GraphBin2 for comparison.
 
 Run `gbintk visualise --help` or `gbintk visualise -h` to list the help message for visualisation.
 
@@ -41,49 +41,53 @@ Options:
 
 The *metaSPAdes* version takes in 4 files as inputs.
 
-* Initial binning resilt (in `.csv` format)
-* Final binning resilt (in `.csv` format)
+* A delimited text file containing the initial binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
+* A delimited text file containing the refined binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
 * Assembly graph file (in `.gfa` format)
 * Contigs file (in `.fasta` format)
 * Contig paths file (in `.paths` format)
-* Abundance file (in `.tsv` format) with a contig in a line and its coverage in each sample separated by tabs.
 
 The *MEGAHIT* version takes in 3 files as inputs.
 
-* Initial binning resilt (in `.csv` format)
-* Final binning resilt (in `.csv` format)
+* A delimited text file containing the initial binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
+* A delimited text file containing the refined binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
 * Assembly graph file (in `.gfa` format)
 * Contigs file (in `.fasta` format)
-* Abundance file (in `.tsv` format) with a contig in a line and its coverage in each sample separated by tabs.
 
 The *Flye* version takes in 4 files as inputs.
 
-* Initial binning resilt (in `.csv` format)
-* Final binning resilt (in `.csv` format)
+* A delimited text file containing the initial binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
+* A delimited text file containing the refined binning result (e.g. `<contig_id>,<bin_number>` in `.csv` format)
 * Assembly graph file (`assembly_graph.gfa`)
 * Contigs file (`assembly.fasta`)
 * Contig paths file (`assembly_info.txt`)
-* Abundance file (in `.tsv` format) with a contig in a line and its coverage in each sample separated by tabs.
 
 ## Example Usage
 
 ```shell
-# SPAdes assembly
-gbintk visualise --assembler spades --initial /path/to/initial_binning_res.csv --final /path/to/final_binning_res.csv --graph /path/to/graph_file.gfa --contigs /path/to/contigs.fasta --paths /path/to/paths_file.paths --output /path/to/output_folder
+# SPAdes assembly available in tests/data/
+gbintk visualise --assembler spades --initial tests/data/5G_metaSPAdes/metacoag_res.csv --final tests/data/5G_metaSPAdes/graphbin_res.csv --graph tests/data/5G_metaSPAdes//assembly_graph_with_scaffolds.gfa --contigs tests/data/5G_metaSPAdes//contigs.fasta --paths tests/data/5G_metaSPAdes/paths_file.paths --output tests/data/5G_metaSPAdes/visualise_results
 
-# MEGAHIT assembly
[…]
