### 2025-04-29 c827be5
DOC: update evaluate docs
