### 2025-04-29 8e1a046
DOC: update graphbin2 docs
