+++ docs/graphbin2_usage.md
@@ -1,5 +1,7 @@
 # Using GraphBin2
 
+A formatted initial binning result from the `prepare` subcommand can be improved by providing it to GraphBin2 using the subcommand `graphbin2`.
+
 Run `gbintk graphbin2 --help` or `gbintk graphbin2 -h` to list the help message for GraphBin2.
 
 ```shell
@@ -40,21 +42,22 @@ The SPAdes version of GraphBin2 takes in 4 files as inputs (required).
 * Contigs file (in `.fasta` format)
 * Assembly graph file (in `.gfa` format)
 * Paths of contigs (in `.paths` format)
-* Binning output from an existing tool (in `.csv` format)
+* A delimited text file containing the initial binning result (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
+* A delimited file containing the contig identifier and its average read coverage for each contig - can be obtained by running a read coverage calculation tool such as [CoverM](https://github.com/wwood/CoverM) or [Koverage](https://github.com/beardymcjohnface/Koverage).
 
 The MEGAHIT version of GraphBin2 takes in 4 files as inputs (required).
 
-* Contigs file (in `.fasta` format)
-* Abundance file (tab separated file with contig ID and coverage in each line)
+* Contigs file (in `.fasta` format)\
 * Assembly graph file (in `.gfa` format)
-* Binning output from an existing tool (in `.csv` format)
+* A delimited text file containing the initial binning result (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
+* A delimited file containing the contig identifier and its average read coverage for each contig
 
 The Flye version of GraphBin2 takes in 4 files as inputs (required).
 
 * Contigs file (in `.fasta` format)
-* Abundance file (tab separated file with contig ID and coverage in each line)
 * Assembly graph file (in `.gfa` format)
-* Binning output from an existing tool (in `.csv` format)
+* A delimited text file containing the initial binning result (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
+* A delimited file containing the contig identifier and its average read coverage for each contig
 
 **Note:** Make sure that the initial binning result consists of contigs belonging to only one bin. GraphBin2 is designed to handle initial contigs which belong to only one bin.
 
@@ -62,19 +65,20 @@ The Flye version of GraphBin2 takes in 4 files as inputs (required).
 ## Example Usage
 
 ```shell
-# SPAdes assembly
-gbintk graphbin2 --assembler spades --contigs /path/to/contigs.fasta --paths /path/to/paths_file.paths --graph /path/to/graph_file.gfa  --binned /path/to/binning_result.csv --abundance /path/to/abundance.tsv --output /path/to/output_folder
+# SPAdes assembly available in tests/data/
+gbintk graphbin2 --assembler spades --graph tests/data/5G_metaSPAdes/graph_file.gfa  --contigs tests/data/5G_metaSPAdes/contigs.fasta --paths tests/data/5G_metaSPAdes/contigs.paths --binned tests/data/5G_metaSPAdes/initial_contig_bins.csv --abundance tests/data/5G_metaSPAdes/abundance.tsv --output tests/data/5G_metaSPAdes/graphbin2_results
 
-# MEGAHIT version
-gbintk graphbin2 --assembler megahit --graph /path/to/final.gfa --contigs /path/to/final.contigs.fa --binned /path/to/binning_result.csv --abundance /path/to/abundance.tsv --output /path/to/output_folder
+# MEGAHIT assembly available in tests/data/
[…]
