+++ docs/graphbin_usage.md
@@ -1,5 +1,7 @@
 # Using GraphBin
 
+A formatted initial binning result from the `prepare` subcommand can be improved by providing it to GraphBin using the subcommand `graphbin`.
+
 Run `gbintk graphbin --help` or `gbintk graphbin -h` to list the help message for GraphBin.
 
 ```shell
@@ -31,49 +33,49 @@ Options:
 
 `max_iteration` and `diff_threshold` parameters are set by default to `100` and `0.1` respectively. However, the user can specify them when running GraphBin.
 
-## Inputs
+## Input Format
 
 For the SPAdes version, GraphBin takes in 3 files as inputs (required).
 
 * Assembly graph file (in `.gfa` format)
-* Contigs file (`contigs.fasta` file in FASTA format)
-* Paths of contigs (`contigs.paths` file)
-* Binning output from an existing tool (in `.csv` format)
+* Contigs file (in `.fasta` format)
+* Contig paths file (`.paths` format)
+* A delimited text file containing the initial binning result (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
 
 For the MEGAHIT version, GraphBin takes in 3 files as inputs (required).
 
-* Assembly graph file (in `.gfa` format. To convert fastg to gfa refer [here](https://github.com/Vini2/GraphBin/blob/master/support/README.md#fastg2gfa))
-* Contigs file (in FASTA format)
-* Binning output from an existing tool (in `.csv` format)
+* Assembly graph file (in `.gfa` format)
+* Contigs file (in `.fasta` format)
+* A delimited text file containing the initial binning result (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
 
 For the Flye version, GraphBin takes in 3 files as inputs (required).
 
-* Assembly graph file (in `.gfa` format)
-* Contigs file (`assembly.fasta` file in FASTA format)
-* Paths of contigs (`assembly_info.txt` file)
-* Binning output from an existing tool (in `.csv` format)
+* Assembly graph file (`assembly_graph.gfa`)
+* Contigs file (`assembly.fasta`)
+* Contig paths file (`assembly_info.txt`)
+* A delimited text file containing the initial binning result (e.g.`<contig_id>,<groud_truth_bin>` in `.csv` format)
 
 **Note:** Make sure that the initial binning result consists of contigs belonging to only one bin. GraphBin is designed to handle initial contigs which belong to only one bin. Multiple bins for the initial contigs are not supported.
 
-**Note:** You can specify the delimiter for the initial binning result file and the final output file using the delimiter paramter. Enter the following values for different delimiters; `,` for a comma, `;` for a semicolon, `$'\t'` for a tab, `" "` for a space and `|` for a pipe.
+**Note:** You can specify the delimiter for the initial binning result file and the final output file using the delimiter paramter. Supports comma (`comma`) and tab (`tab`).
 
 ## Example Usage
 
 ```shell
-# SPAdes assembly
-gbintk graphbin --assembler spades --graph /path/to/graph_file.gfa --contigs /path/to/contigs.fasta --paths /path/to/paths_file.paths --binned /path/to/binning_result.csv --output /path/to/output_folder
+# SPAdes assembly available in tests/data/
+gbintk graphbin --assembler spades --graph tests/data/5G_metaSPAdes/assembly_graph_with_scaffolds.gfa --contigs tests/data/5G_metaSPAdes/contigs.fasta --paths tests/data/5G_metaSPAdes/contigs.paths --binned tests/data/5G_metaSPAdes/initial_contig_bins.csv --output tests/data/5G_metaSPAdes/graphbin_results
 
[…]
