### 2025-04-29 e17dd3c
DOC: update graphbin2 docs
