### 2025-04-29 df317e6
DOC: update graphbin/graphbin2 docs
