+++ paper/paper.bib
@@ -486,9 +486,11 @@ url={https://doi.org/10.1186/s13059-022-02832-6}
 }
 
 @misc{deepthought_hpc,
-  title        = {DeepThought HPC},
-  author       = {{Flinders University}},
-  year         = {2024},
-  howpublished = {\url{https://doi.org/10.25957/FLINDERS.HPC.DEEPTHOUGHT}},
-  note         = {High Performance Computing resource at Flinders University},
+  doi = {10.25957/FLINDERS.HPC.DEEPTHOUGHT},
+  url = {https://deepthoughtdocs.flinders.edu.au/en/latest/},
+  author = {{Flinders University}},
+  keywords = {HPC, High performance computing},
+  title = {Deep Thought (HPC)},
+  publisher = {Flinders University},
+  year = {2021}
 }
