### 2025-04-30 4c92476
DOC: update citations
