### 2025-04-30 8f72558
DOC: add dependencies in docs
