### 2025-04-30 metagentools/gbintk#15 @Vini2
Hello @telatin!

Thanks for taking the time to review our paper. To respond to your comments:

1. I have removed the usage instructions from the paper and summarised the functions in a separate section called [*Functionality*](https://github.com/metagentools/gbintk/blob/JOSS/paper/paper.md#functionality). The usage instructions are included in the documentation hosted at [Read the Docs](https://gbintk.readthedocs.io/en/latest/). I have also cleaned up the [*Statement of need*](https://github.com/metagentools/gbintk/blob/JOSS/paper/paper.md#statement-of-need) section with more background on the tool and how it helps researchers focus on the actual analysis rather than dealing with the technical challenges of using multiple tools.
2. Each subcommand [`graphbin`](https://gbintk.readthedocs.io/en/latest/graphbin_preprocess/), [`graphbin2`](https://gbintk.readthedocs.io/en/latest/graphbin2_preprocess/) and [`metacoag`](https://gbintk.readthedocs.io/en/latest/metacoag_preprocess/) has a section named *Preprocess* that explains the upstream steps required.

Please have a look at these changes and let me know if further changes are needed.

Thanks!
