### 2025-05-01 metagentools/gbintk#16 @Vini2
Hello @cumbof!

Thanks for taking the time to review our paper. To respond to your comments:

1. I have added a checklist of all the dependencies and external software in the install section of the [documentation](https://gbintk.readthedocs.io/en/latest/install/).
2. The Python version and all dependencies are listed in the install section of the [documentation](https://gbintk.readthedocs.io/en/latest/install/).
3. I have updated the command line examples with real paths based on the data available in [tests](https://github.com/metagentools/gbintk/tree/main/tests).
4. I have hosted all the example data on [Zenodo](https://zenodo.org/records/15313645) and the link has been added to the paper.
5. I have corrected the caption of Figure 3.

Please have a look at these changes and let me know if I need to make more changes.

Thanks!
