+++ src/gbintk/support/prep_result.py
@@ -21,7 +21,7 @@ __author__ = "Vijini Mallawaarachchi"
 __copyright__ = "Copyright 2023-2024, GraphBin-Tk Project"
 __credits__ = ["Vijini Mallawaarachchi", "Anuradha Wickramarachchi", "Yu Lin"]
 __license__ = "GPL-3.0"
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 __maintainer__ = "Vijini Mallawaarachchi"
 __email__ = "viji.mallawaarachchi@gmail.com"
 __status__ = "Production/Stable"
