### 2025-05-01 metagentools/gbintk#16 @cumbof
Thanks @Vini2

I don’t have any other comments, so I’m going to close this issue.
Well done.
