### 2025-05-02 844624f
DOC: update paper citations
