+++ paper/paper.bib
@@ -465,6 +465,23 @@ doi={10.1038/s41564-018-0171-1},
 url={https://doi.org/10.1038/s41564-018-0171-1}
 }
 
+@Article{Kariin:1995,
+  author={Kariin, Samuel
+  and Burge, Chris},
+  title={Dinucleotide relative abundance extremes: a genomic signature},
+  journal={Trends in Genetics},
+  year={1995},
+  month={Jul},
+  day={01},
+  publisher={Elsevier},
+  volume={11},
+  number={7},
+  pages={283-290},
+  issn={0168-9525},
+  doi={10.1016/S0168-9525(00)89076-9},
+  url={https://doi.org/10.1016/S0168-9525(00)89076-9}
+}
+
 @Article{Wang:2023,
 author={Wang, Ziye
 and Huang, Pingqin
