### 2025-05-05 metagentools/gbintk#14 @Vini2
Marked as complete [here](https://github.com/openjournals/joss-reviews/issues/7713#issuecomment-2843989975). Closing the issue.
