### 2025-05-12 #7713 @arfon
> Yes, to me the main (suggested) changes are in the paper itself (https://github.com/metagentools/gbintk/issues/15#issuecomment-2777885989)

We'd like to credit for your review effort here. Could you please return to your checklist (linked to from the top of the issue) and complete it when you have a moment.

Thanks!
