### 2025-05-15 metagentools/gbintk#15 @Vini2
Hello @telatin,

Could you please take a look at this comment?
https://github.com/openjournals/joss-reviews/issues/7713#issuecomment-2870922222

Thanks!
