### 2025-05-20 metagentools/gbintk#15 @telatin
Well done @Vini2  :)
