+++ src/gbintk/support/evaluate.py
@@ -14,7 +14,7 @@ import scipy.special
 from tabulate import tabulate
 
 __author__ = "Vijini Mallawaarachchi"
-__copyright__ = "Copyright 2023-2024, GraphBin-Tk Project"
+__copyright__ = "Copyright 2023-2025, GraphBin-Tk Project"
 __credits__ = ["Vijini Mallawaarachchi", "Anuradha Wickramarachchi", "Yu Lin"]
 __license__ = "GPL-3.0"
 __version__ = "1.0.2"
