+++ src/gbintk/support/visualise_result_SPAdes.py
@@ -22,7 +22,7 @@ from graphbin.bidirectionalmap.bidirectionalmap import BidirectionalMap
 from igraph import *
 
 __author__ = "Vijini Mallawaarachchi"
-__copyright__ = "Copyright 2023-2024, GraphBin-Tk Project"
+__copyright__ = "Copyright 2023-2025, GraphBin-Tk Project"
 __credits__ = ["Vijini Mallawaarachchi", "Anuradha Wickramarachchi", "Yu Lin"]
 __license__ = "GPL-3.0"
 __version__ = "1.0.2"
