+++ docs/evaluate.md
@@ -50,7 +50,7 @@ gbintk evaluate --binned Sim-20G/graphbin_output.csv --groundtruth Sim-20G/groun
 
 You will get a file named `evaluation_results.txt` that contains the $K \times S$ matrix and the calculated evaluation metrics.
 
-**Note:** Make sure that the binning result consists of contigs belonging to only one bin. The evaluation metrics consider contigs which belong to only one bin. Please refer to the the [GraphBin2 publication](https://doi.org/10.1186/s13015-021-00185-6) for details on metrics used to evaluate multi-labelled contigs.
+**Note:** Make sure that the binning result consists of contigs belonging to only one bin. The evaluation metrics consider contigs which belong to only one bin. Please refer to the [GraphBin2 publication](https://doi.org/10.1186/s13015-021-00185-6) for details on metrics used to evaluate multi-labelled contigs.
 
 ### Plotting Evaluation Results
 
