+++ docs/index.md
@@ -1,11 +1,11 @@
 # GraphBin-Tk: assembly graph-based metagenomic binning toolkit
 
-GraphBin-Tk combines assembly graph-based metagenomic bin-refinement and binning techniques [GraphBin](https://github.com/metagentools/GraphBin), [GraphBin2](https://github.com/metagentools/GraphBin2) and [MetaCoAG](https://github.com/metagentools/MetaCoAG) along with additional processing functionality to visualise and evaluate results, into one comprehensive toolkit.
+GraphBin-Tk combines the assembly graph-based metagenomic bin-refinement and binning techniques of [GraphBin](https://github.com/metagentools/GraphBin), [GraphBin2](https://github.com/metagentools/GraphBin2) and [MetaCoAG](https://github.com/metagentools/MetaCoAG) along with additional processing functionalities to visualise and evaluate results, into one comprehensive toolkit.
 
 ![](images/gbintk_workflow.png)
 
 **Why GraphBin-Tk?**
 
-The assembly graph-based metagenomic binning tools GraphBin, GraphBin2 and MetaCoAG exist as individual software and running them individually can be complex, time-consuming and less accessible. These tools require different dependencies and rely on additional software to function properly. Users have to manually setup each tool, manage different output formats and write custom code to perform additional analysis, which can raise the risk of errors. Furthermore, existing metagenomic binning toolkits and wrappers do not include these graph-based binning tools.
+The assembly graph-based metagenomic binning tools GraphBin, GraphBin2 and MetaCoAG exist as individual software and running them individually can be complex, time-consuming and less accessible. These tools require different dependencies and rely on additional software to function properly. Users have to manually set up each tool, manage different output formats and write custom code to perform additional analysis, which can raise the risk of errors. Furthermore, existing metagenomic binning toolkits and wrappers do not include these graph-based binning tools.
 
 GraphBin-Tk integrates GraphBin, GraphBin2 and MetaCoAG with additional functions such as visualisation and evaluation, into a comprehensive toolkit. This enables users to seamlessly perform a wide range of tasks related to metagenomic binning and eliminates any compatibility issues that may arise from running separate tools.
