### 2025-05-27 c7d7f4e
MAINT: fix typos in visualise
