### 2025-05-27 96bbcc5
DOC: Create CITATION.cff
