+++ CITATION.cff
@@ -1,0 +1,24 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+title: "GraphBin-Tk: assembly graph-based metagenomic binning toolkit"
+authors:
+- family-names: "Mallawaarachchi"
+  given-names: "Vijini"
+  orcid: "https://orcid.org/0000-0002-2651-8719"
+- family-names: "Wickramarachchi"
+  given-names: "Anuradha"
+  orcid: "https://orcid.org/0000-0003-4160-5965"
+- family-names: "McArthur"
+  given-names: "Robert"
+  orcid: "https://orcid.org/0000-0001-9099-339X"
+- family-names: "Lang"
+  given-names: "Yapeng"
+  orcid: "https://orcid.org/0009-0001-6828-0964"
+- family-names: "Caley"
+  given-names: "Katherine"
+  orcid: "https://orcid.org/0000-0002-8459-6168"
+- family-names: "Huttley"
+  given-names: "Gavin"
+  orcid: "https://orcid.org/0000-0001-7224-2074"
+doi: 10.5281/zenodo.15524165
+url: "https://github.com/metagentools/gbintk"
