### 2025-05-28 #7713 @Vini2
Hi @arfon!

Thanks for the information. I've created a new release, proofread the paper and created the Zenodo archive. The DOI is [https://doi.org/10.5281/zenodo.15531680](https://doi.org/10.5281/zenodo.15531680).
